// backend/src/skill/skillCategory.enum.ts
// src/skill/skill-category.enum.ts
export enum SkillCategory {
  LANGUAGE = 'Language',
  FRAMEWORK = 'Framework',
  DATABASE = 'Database',
  DEVOPS = 'DevOps',
  CLOUD = 'Cloud',
  AI_ML = 'AI/ML',
  DATA = 'Data',
  CYBERSECURITY = 'Cybersecurity',
  MOBILE = 'Mobile',
  EMBEDDED = 'Embedded',
  NETWORK = 'Network',
  TESTING = 'Testing',
  SOFT_SKILL = 'Soft Skill',
}